ErrorService moved from Controller component to FlashMessage component
- component is part of FlashMessage and is used only with FlashMessage as a complex
